requirements.txt
openai-whisper>=20231117
python-docx>=1.1.0
numpy>=1.26.0
